Removed old API stuff
