+++ dwitter/api/models.py
@@ -1,0 +1,4 @@
+from django.db import models
+
+class APIDweet(models.Model):
+    test = "hoi"
