+++ dwitter/permissions.py
@@ -1,7 +1,4 @@
-from rest_framework import permissions
-
-
-class IsAuthorOrReadOnly(permissions.BasePermission):
+class IsAuthorOrReadOnly():#permissions.BasePermission
     """
     Custom permission to only allow owners of an object to edit it.
     """
@@ -9,7 +6,7 @@ class IsAuthorOrReadOnly(permissions.BasePermission):
     def has_object_permission(self, request, view, obj):
         # Read permissions are allowed to any request,
         # so we'll always allow GET, HEAD or OPTIONS requests.
-        if request.method in permissions.SAFE_METHODS:
+        if request.method in ["GET","HEAD","OPTIONS"]: #permissions.SAFE_METHODS
             return True
 
         # Write permissions are only allowed to the owner of the snippet.
