+++ requirements.txt
@@ -8,7 +8,6 @@ django-filter==1.1
 django-hosts==3.0
 django-registration-redux==2.4
 django-storages-redux==1.3.2
-django==1.11.15
-djangorestframework==3.7.1
+django==1.11.17
 flake8==2.6.2
 newrelic==4.2.0.100
